+++ packages/sw/src/components/ManifestLink.tsx
@@ -1,5 +1,12 @@
+import type { LinkHTMLAttributes } from 'react';
 import React from 'react';
 
-export const ManifestLink = ({ manifestUrl = '/manifest.webmanifest' }: { manifestUrl?: string }) => {
-  return <link rel="manifest" href={manifestUrl} />;
+export const ManifestLink = ({
+  crossOrigin,
+  manifestUrl = '/manifest.webmanifest',
+}: {
+  manifestUrl?: string;
+  crossOrigin?: LinkHTMLAttributes<HTMLLinkElement>['crossOrigin'];
+}) => {
+  return <link rel="webmanifest" href={manifestUrl} crossOrigin={crossOrigin} />;
 };