+++ packages/sw/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @remix-pwa/sw 3.0.10-dev.1 (2024-09-24)
+
+
+### Bug Fixes
+
+* **sw:** changed manifest link `rel` from `manifest` to `webmanifest` 942f76a
+
 ## @remix-pwa/sw 3.0.9 (2024-08-11)